pyproject.toml
[tool.poetry]
name = "customer_service"
version = "0.1.0"
description = ""
authors = ["mehdi <13610522>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
djangogrpcframework = "^0.2.1"
grpcio = "^1.48.0"
grpcio-tools = "^1.48.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.3"
ipython = "^8.4.0"
ipdb = "^0.13.9"
celery = "^5.2.7"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
